+++ app/views/DetailsPanel.h
@@ -11,7 +11,7 @@
 #include "LabelImageCache.h"
 #include "model/ScopedEventSink.h"
 
-#include <ctb/tasks/tasks.h>
+#include <ctb/task/tasks.h>
 #include <wx/panel.h>
 #include <wx/timer.h>
